+++ src/types/Sandbox.ts
@@ -3,7 +3,7 @@ import SocketClient from "./SocketClient";
 /** @noSelf */
 export default interface Sandbox {
   isSocketInitialized(): boolean;
-  connectLocalhost(port: int): SocketClient;
+  connectLocalhost(port: int, useTCP: boolean): SocketClient;
   traceback(): void;
   getParentFunctionDescription(levels: int): string;
 }